Fix score and interactions

+++ Assets/Scenes/Test/Iley/ScoreManager.cs
@@ -3,7 +3,7 @@ using UnityEngine.UI;
 
 public class ScoreManager : MonoBehaviour {
     public Text scoreText;
-    public static int scoreCount;
+    public static float scoreCount;
 
     public void Update() {
         scoreText.text = "Score: " + Mathf.Round(scoreCount);
